rename CourseCreate.tsx -> SearchCourse.tsx
