+++ frontend/src/App.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { BrowserRouter, Route, Routes, Navigate } from 'react-router-dom';
 
 import './App.css';
-import CourseCreate from './containers/CourseCreate/CourseCreate';
 import PostCourse from './containers/CourseCreate/PostCourse';
+import SearchCourse from './containers/CourseCreate/SearchCourse';
 import CourseDetail from './containers/CourseDetail/DetailPage';
 import CourseList from './containers/CourseList/CourseList';
 import MainPage from './containers/MainPage/MainPage';
@@ -15,7 +15,7 @@ function App() {
         <Routes>
           <Route path="/courses" element={<CourseList />} />
           <Route path="/course/:id" element={<CourseDetail />} />
-          <Route path="/course-create/search" element={<CourseCreate />} />
+          <Route path="/course-create/search" element={<SearchCourse />} />
           <Route path="/course-create/post" element={<PostCourse />} />
           {/* <Route path="/register/" element={<Register />} /> */}
           <Route path="/" element={<MainPage />} />
