+++ frontend/src/components/SearchBar/SearchBar.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 
-import { MarkerProps } from '../../containers/CourseCreate/CourseCreate';
-import styles from '../../containers/CourseCreate/CourseCreate.module.scss';
+import { MarkerProps } from '../../containers/CourseCreate/SearchCourse';
+import styles from '../../containers/CourseCreate/SearchCourse.module.scss';
 
 type SearchProps = {
   markers: MarkerProps[];
