+++ frontend/src/store/slices/course.test.ts
@@ -2,7 +2,7 @@ import { AnyAction, configureStore, EnhancedStore } from '@reduxjs/toolkit';
 import axios from 'axios';
 import { ThunkMiddleware } from 'redux-thunk';
 
-import { DataProps, FeatureProps, MarkerProps } from '../../containers/CourseCreate/CourseCreate';
+import { DataProps, FeatureProps, MarkerProps } from '../../containers/CourseCreate/SearchCourse';
 import reducer, {
   CourseState,
   CourseType,
